# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserDictWord": {
        "description": "辞書のコンパイルに使われる情報",
        "properties": {
          "accent_associative_rule": {
            "title": "アクセント結合規則",
            "type": "string"
          },
          "accent_type": {
            "title": "アクセント型",
            "type": "integer"
          },
          "context_id": {
            "default": 1348,
            "title": "文脈ID",
            "type": "integer"
          },
          "inflectional_form": {
            "title": "活用形",
            "type": "string"
          },
          "inflectional_type": {
            "title": "活用型",
            "type": "string"
          },
          "mora_count": {
            "title": "モーラ数",
            "type": "integer"
          },
          "part_of_speech": {
            "title": "品詞",
            "type": "string"
          },
          "part_of_speech_detail_1": {
            "title": "品詞細分類1",
            "type": "string"
          },
          "part_of_speech_detail_2": {
            "title": "品詞細分類2",
            "type": "string"
          },
          "part_of_speech_detail_3": {
            "title": "品詞細分類3",
            "type": "string"
          },
          "priority": {
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "優先度",
            "type": "integer"
          },
          "pronunciation": {
            "title": "発音",
            "type": "string"
          },
          "stem": {
            "title": "原形",
            "type": "string"
          },
          "surface": {
            "title": "表層形",
            "type": "string"
          },
          "yomi": {
            "title": "読み",
            "type": "string"
          }
        },
        "required": [
          "surface",
          "priority",
          "part_of_speech",
          "part_of_speech_detail_1",
          "part_of_speech_detail_2",
          "part_of_speech_detail_3",
          "inflectional_type",
          "inflectional_form",
          "stem",
          "yomi",
          "pronunciation",
          "accent_type",
          "accent_associative_rule"
        ],
        "title": "UserDictWord",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WordTypes": {
        "description": "fastapiでword_type引数を検証する時に使用するクラス",
        "enum": [
          "PROPER_NOUN",
          "COMMON_NOUN",
          "VERB",
          "ADJECTIVE",
          "SUFFIX"
        ],
        "title": "WordTypes",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "DUMMY の音声合成エンジンです。",
    "title": "DUMMY Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/import_user_dict": {
      "post": {
        "description": "他のユーザー辞書をインポートします。",
        "operationId": "import_user_dict_words_import_user_dict_post",
        "parameters": [
          {
            "description": "重複したエントリがあった場合、上書きするかどうか",
            "in": "query",
            "name": "override",
            "required": true,
            "schema": {
              "description": "重複したエントリがあった場合、上書きするかどうか",
              "title": "Override",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "$ref": "#/components/schemas/UserDictWord"
                },
                "description": "インポートするユーザー辞書のデータ",
                "title": "Import Dict Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import User Dict Words",
        "tags": [
          "ユーザー辞書"
        ]
      }
    },
...
    "/user_dict_word": {
      "post": {
        "description": "ユーザー辞書に言葉を追加します。",
        "operationId": "add_user_dict_word_user_dict_word_post",
        "parameters": [
          {
            "description": "言葉の表層形",
            "in": "query",
            "name": "surface",
            "required": true,
            "schema": {
              "description": "言葉の表層形",
              "title": "Surface",
              "type": "string"
            }
          },
          {
            "description": "言葉の発音（カタカナ）",
            "in": "query",
            "name": "pronunciation",
            "required": true,
            "schema": {
              "description": "言葉の発音（カタカナ）",
              "title": "Pronunciation",
              "type": "string"
            }
          },
          {
            "description": "アクセント型（音が下がる場所を指す）",
            "in": "query",
            "name": "accent_type",
            "required": true,
            "schema": {
              "description": "アクセント型（音が下がる場所を指す）",
              "title": "Accent Type",
              "type": "integer"
            }
          },
          {
            "description": "PROPER_NOUN（固有名詞）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか",
            "in": "query",
            "name": "word_type",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/WordTypes"
                }
              ],
              "description": "PROPER_NOUN（固有名詞）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか"
            }
          },
          {
            "description": "単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨",
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "description": "単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨",
              "maximum": 10.0,
              "minimum": 0.0,
              "title": "Priority",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Add User Dict Word User Dict Word Post",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add User Dict Word",
        "tags": [
          "ユーザー辞書"
        ]
      }
    },
    "/user_dict_word/{word_uuid}": {
      "delete": {
        "description": "ユーザー辞書に登録されている言葉を削除します。",
        "operationId": "delete_user_dict_word_user_dict_word__word_uuid__delete",
        "parameters": [
          {
            "description": "削除する言葉のUUID",
            "in": "path",
            "name": "word_uuid",
            "required": true,
            "schema": {
              "description": "削除する言葉のUUID",
              "title": "Word Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete User Dict Word",
        "tags": [
          "ユーザー辞書"
        ]
      },
      "put": {
        "description": "ユーザー辞書に登録されている言葉を更新します。",
        "operationId": "rewrite_user_dict_word_user_dict_word__word_uuid__put",
        "parameters": [
          {
            "description": "更新する言葉のUUID",
            "in": "path",
            "name": "word_uuid",
            "required": true,
            "schema": {
              "description": "更新する言葉のUUID",
              "title": "Word Uuid",
              "type": "string"
            }
          },
          {
            "description": "言葉の表層形",
            "in": "query",
            "name": "surface",
            "required": true,
            "schema": {
              "description": "言葉の表層形",
              "title": "Surface",
              "type": "string"
            }
          },
          {
            "description": "言葉の発音（カタカナ）",
            "in": "query",
            "name": "pronunciation",
            "required": true,
            "schema": {
              "description": "言葉の発音（カタカナ）",
              "title": "Pronunciation",
              "type": "string"
            }
          },
          {
            "description": "アクセント型（音が下がる場所を指す）",
            "in": "query",
            "name": "accent_type",
            "required": true,
            "schema": {
              "description": "アクセント型（音が下がる場所を指す）",
              "title": "Accent Type",
              "type": "integer"
            }
          },
          {
            "description": "PROPER_NOUN（固有名詞）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか",
            "in": "query",
            "name": "word_type",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/WordTypes"
                }
              ],
              "description": "PROPER_NOUN（固有名詞）、COMMON_NOUN（普通名詞）、VERB（動詞）、ADJECTIVE（形容詞）、SUFFIX（語尾）のいずれか"
            }
          },
          {
            "description": "単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨。",
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "description": "単語の優先度（0から10までの整数）。数字が大きいほど優先度が高くなる。1から9までの値を指定することを推奨。",
              "maximum": 10.0,
              "minimum": 0.0,
              "title": "Priority",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rewrite User Dict Word",
        "tags": [
          "ユーザー辞書"
        ]
      }
    },
...
  }
}